+++ boat/prima/src/prima.ts
@@ -1102,7 +1102,7 @@ export class Prima {
       aria: result.ariaSnapshot,
       html: await result.combinedHtml(),
       screenshot: result.screenshot,
-      requests: this.bot.requestStore().getMadeRequests(),
+      requests: this.bot.requestStore().getCapturedRequests(),
     });
   }